feat: Create separate entrypoint for gosling.js/utils
